moved data processing functions from app.py to preprocess.py
